+++ tests/conftest.py
@@ -1,6 +1,8 @@
 """
 PyTest configuration
 """
+import os
+
 from typing import List
 
 import numpy as np
@@ -33,10 +35,44 @@ from baybe.strategies.sampling import RandomRecommender
 from baybe.strategies.strategy import Strategy
 from baybe.targets import NumericalTarget, Objective
 
-
 # All fixture functions have prefix 'fixture_' and explicitly declared name so they
 # can be reused by other fixtures, see
 # https://docs.pytest.org/en/stable/reference/reference.html#pytest-fixture
+
+
+@pytest.fixture(scope="session", autouse=True)
+def disable_telemetry():
+    """
+    Disables telemetry during pytesting via fixture
+    """
+    # Remember the original value of the environment variables
+    telemetry_enabled_before = os.environ.get("BAYBE_TELEMETRY_ENABLED")
+    telemetry_userhash_before = os.environ.get("BAYBE_DEBUG_FAKE_USERHASH")
+    telemetry_hosthash_before = os.environ.get("BAYBE_DEBUG_FAKE_HOSTHASH")
+
+    # Set the environment variable to a certain value for the duration of the tests
+    os.environ["BAYBE_TELEMETRY_ENABLED"] = "false"
+    os.environ["BAYBE_DEBUG_FAKE_USERHASH"] = "PYTEST"
+    os.environ["BAYBE_DEBUG_FAKE_HOSTHASH"] = "PYTEST"
+
+    # Yield control to the tests
+    yield
+
+    # Restore the original value of the environment variables
+    if telemetry_enabled_before is not None:
+        os.environ["BAYBE_TELEMETRY_ENABLED"] = telemetry_enabled_before
+    else:
+        os.environ.pop("BAYBE_TELEMETRY_ENABLED")
+
+    if telemetry_userhash_before is not None:
+        os.environ["BAYBE_DEBUG_FAKE_USERHASH"] = telemetry_userhash_before
+    else:
+        os.environ.pop("BAYBE_DEBUG_FAKE_USERHASH")
+
+    if telemetry_hosthash_before is not None:
+        os.environ["BAYBE_DEBUG_FAKE_HOSTHASH"] = telemetry_hosthash_before
+    else:
+        os.environ.pop("BAYBE_DEBUG_FAKE_HOSTHASH")
 
 
 # Add option to only run fast tests