+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Activated mypy for the first few modules and fixed their type issues
 - Automatic (de-)serialization and `SerialMixin` class
 - Basic serialization example and tests
+- Telemetry via OpenTelemetry
 
 ### Changed
 - Switched from pydantic to attrs in all modules except constraints.py